tests/requirements.txt
boto3==1.33.10
botocore==1.33.10
pytest==8.0.0
psycopg2-binary==2.9.9
